Refactoring: card style 코드분리
