+++ components/common/scheduleCard/CardImage.tsx
@@ -1,32 +1,10 @@
-import { ContentsDataType } from '@/types/inSheet';
-import Image from 'next/image';
-import { MouseEvent, useCallback, useRef, useState } from 'react';
+import { ImageBox } from '@/components/common/scheduleCard/Style';
 import { DEFAULT_BLUR_BASE64 } from '@/consts';
 import altImage from '@/images/thumbnail_alt_img.png';
-import styled from '@emotion/styled';
+import { ContentsDataType } from '@/types/inSheet';
 import { gtagClickAtag } from '@/utils/gtag';
-import { WIDTH } from '@/styles/var';
-
-const ImageBox = styled.a`
-  position: relative;
-  box-sizing: border-box;
-  margin: auto;
-  width: 12rem;
-  aspect-ratio: 16 / 9;
-  border-radius: 5px;
-  height: 100%;
-  overflow: hidden;
-  background-color: #fff;
-
-  & > img {
-    object-fit: cover;
-  }
-
-  @media (min-width: ${WIDTH.SM}) {
-    width: 100%;
-    flex-basis: 200%;
-  }
-`;
+import Image from 'next/image';
+import { MouseEvent, useCallback, useRef, useState } from 'react';
 
 interface ScheduleCardImageProps {
   content: ContentsDataType;
