+++ components/common/scheduleCard/CardPlaceHolder.tsx
@@ -1,9 +1,4 @@
-import { cardBase } from '@/components/common/scheduleCard/Card';
-import styled from '@emotion/styled';
-
-const Card = styled.div`
-  ${cardBase}
-`;
+import { cardBase } from '@/components/common/scheduleCard/Style';
 
 const ITEMS = Array.from({ length: 10 }, (_, i) => i);
 
@@ -11,7 +6,7 @@ const ScheduleCardPlaceHolder = () => {
   return (
     <>
       {ITEMS.map((index) => (
-        <Card key={`schedule_card_placeHolder_${index}`} />
+        <div key={`schedule_card_placeHolder_${index}`} css={cardBase} />
       ))}
     </>
   );
