+++ components/common/scheduleCard/Style.tsx
@@ -1,0 +1,124 @@
+import { BEZIER_CURVE, boxShadow, COLORS, textLine, WIDTH } from '@/styles/var';
+import { css } from '@emotion/react';
+import styled from '@emotion/styled';
+
+export const cardBase = css`
+  margin: 0;
+  flex: 1 0 auto;
+  width: 15rem;
+  max-width: 100%;
+
+  @media (min-width: ${WIDTH.SM}) {
+    max-width: 50%;
+  }
+`;
+
+export const Card = styled.div`
+  ${cardBase}
+  ${boxShadow}
+  display: flex;
+  padding: 0.25rem;
+  gap: 0.25rem;
+  background-color: #fff;
+  border-radius: 5px;
+  transition: all 0.2s ${BEZIER_CURVE};
+
+  &:hover {
+    transform: scale(1.02);
+  }
+
+  @media (min-width: ${WIDTH.SM}) {
+    flex-direction: column;
+    padding: 0.5rem;
+    gap: 0.5rem;
+  }
+
+  &.closed {
+    background-color: ${COLORS['light-blue']};
+  }
+
+  &.stream {
+    background-color: ${COLORS['light-yellow']};
+  }
+`;
+
+export const DescBox = styled.div`
+  margin-bottom: 0.2rem;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-flow: column;
+  gap: 0.5rem;
+
+  .channelName {
+    font-weight: 500;
+    font-size: 1.2rem;
+
+    &.stream {
+      font-weight: 600;
+      color: ${COLORS['highlight-font']};
+    }
+  }
+
+  .title {
+    color: ${COLORS.font};
+    ${textLine(2, 1.25)}
+
+    &.stream {
+      font-weight: 500;
+      color: ${COLORS.secondary};
+    }
+  }
+
+  .time {
+    display: flex;
+    font-size: 0.6rem;
+    margin-bottom: 0.2rem;
+    align-items: center;
+    gap: 0.5rem;
+  }
+
+  .link {
+    display: flex;
+    gap: 0.5rem;
+    font-size: 0.75rem;
+    justify-content: flex-end;
+    margin-right: 0.5rem;
+
+    & > button {
+      font-size: inherit;
+      font-weight: 600;
+      color: ${COLORS.salmon};
+    }
+  }
+`;
+
+export const ImageBox = styled.a`
+  position: relative;
+  box-sizing: border-box;
+  margin: auto;
+  width: 12rem;
+  aspect-ratio: 16 / 9;
+  border-radius: 5px;
+  height: 100%;
+  overflow: hidden;
+  background-color: #fff;
+
+  & > img {
+    object-fit: cover;
+  }
+
+  @media (min-width: ${WIDTH.SM}) {
+    width: 100%;
+    flex-basis: 200%;
+  }
+`;
+
+export const StatusBox = styled.div`
+  display: flex;
+  align-items: center;
+  font-weight: 500;
+  gap: 0.125rem;
+  color: ${COLORS['highlight-font']};
+  vertical-align: middle;
+`;
